Added resource based authorization

+++ Restaurants.Application/Dishes/Command/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Command.DeleteDishes;
@@ -9,7 +11,8 @@ namespace Restaurants.Application.Dishes.Command.DeleteDishes;
 public class DeleteDishesForRestaurantCommandHandler(
     ILogger<DeleteDishesForRestaurantCommandHandler> logger,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesForRestaurantCommand>
+    IDishesRepository dishesRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
     public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -17,6 +20,8 @@ public class DeleteDishesForRestaurantCommandHandler(
         var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
         if (restaurant is null)
             throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperaion.Delete))
+            throw new ForbidException();
         await dishesRepository.DeleteAllDishesAsync(restaurant.Dishes);
     }
 }
